Simulate Daily Virus Incidence Data

Generates simulated daily incidence data for specified respiratory viruses over a defined number of days. Each virus is modeled using a Gaussian-like curve, parameterized by peak day, amplitude, and scale.

Usage

simulate_data(
  days = 365,
  peaks = c(flu_a = 90, rsv = 110, sars_cov2 = 160),
  amplitudes = c(flu_a = 50, rsv = 40, sars_cov2 = 20),
  scales = c(flu_a = -0.004, rsv = -0.005, sars_cov2 = -0.001),
  time_offset = 0,
  noise_sd = 5,
  start_date = "2024-01-07"
)

Arguments

days

Integer. Number of days to simulate (default is 365).

peaks

Named numeric vector. Peak day for each virus (e.g., c("flua"=90,"rsv"=110,"sars_cov2"=160)).

amplitudes

Named numeric vector. Amplitude for each virus's peak (e.g., c("flua"=50,"rsv"=40,"sars_cov2"=20)).

scales

Named numeric vector. Scale controlling spread of the peak for each virus (e.g., c("flua"=-0.004,"rsv"=-0.005,"sars_cov2"=-0.001)).

time_offset

Integer. Number of days to offset start of the simulation. useful if want to test data with larger values in the middle of a respiratory season.

noise_sd

numeric or named numeric. Gaussian noise applied to each virus signal. can either be a single value or named for each virus e.g., c("flu_a"=2,"rsv"=5,"sars_cov2"=7)

start_date

string

Value

A data frame with daily simulated incidence counts for each virus, including a date column.

Examples

simulate_data()
#>           date flu_a rsv sars_cov2
#> 1   2024-01-07     4   0         0
#> 2   2024-01-08     6   1         0
#> 3   2024-01-09     3   0         6
#> 4   2024-01-10     0   0         0
#> 5   2024-01-11     3   0         0
#> 6   2024-01-12     4   0         0
#> 7   2024-01-13     0   0         0
#> 8   2024-01-14     5   4         0
#> 9   2024-01-15     8   0         3
#> 10  2024-01-16     2   0         0
#> 11  2024-01-17     0   1         0
#> 12  2024-01-18     0   1         4
#> 13  2024-01-19     7   0         2
#> 14  2024-01-20     1   9         0
#> 15  2024-01-21     0   2         5
#> 16  2024-01-22     0   0         3
#> 17  2024-01-23     2   1         1
#> 18  2024-01-24     0   0         0
#> 19  2024-01-25     2   0         0
#> 20  2024-01-26     0   1         0
#> 21  2024-01-27     6   2         0
#> 22  2024-01-28     9   1         0
#> 23  2024-01-29     0   8         0
#> 24  2024-01-30     0   0         0
#> 25  2024-01-31     7   0         0
#> 26  2024-02-01     0   0         0
#> 27  2024-02-02     1   0         0
#> 28  2024-02-03     0   7         0
#> 29  2024-02-04     7   0         1
#> 30  2024-02-05     0   0         0
#> 31  2024-02-06     4   0         6
#> 32  2024-02-07     0   5         4
#> 33  2024-02-08     0   4         2
#> 34  2024-02-09     0   0         3
#> 35  2024-02-10     0   0         1
#> 36  2024-02-11     0   3         1
#> 37  2024-02-12     0   0         3
#> 38  2024-02-13     5   1         0
#> 39  2024-02-14     0   0         0
#> 40  2024-02-15     3   0         0
#> 41  2024-02-16     3   5         0
#> 42  2024-02-17     6   0         0
#> 43  2024-02-18     3   4         0
#> 44  2024-02-19     0   3         3
#> 45  2024-02-20     0   3         5
#> 46  2024-02-21     2   0         8
#> 47  2024-02-22     0   0         4
#> 48  2024-02-23     0   0         0
#> 49  2024-02-24     1   5         0
#> 50  2024-02-25     0   0         0
#> 51  2024-02-26     0   0         0
#> 52  2024-02-27     0   3         0
#> 53  2024-02-28     0   0         0
#> 54  2024-02-29     8   0         0
#> 55  2024-03-01     0   7         0
#> 56  2024-03-02     0   0         0
#> 57  2024-03-03     0   0         0
#> 58  2024-03-04     3   4         0
#> 59  2024-03-05     0   0         0
#> 60  2024-03-06     4   0         0
#> 61  2024-03-07     9   0         0
#> 62  2024-03-08     0   0         2
#> 63  2024-03-09     4   1         0
#> 64  2024-03-10     1   5         0
#> 65  2024-03-11    11   0         0
#> 66  2024-03-12     1   0         3
#> 67  2024-03-13     0   5         0
#> 68  2024-03-14     4   8         0
#> 69  2024-03-15     0   3         0
#> 70  2024-03-16     6   0         4
#> 71  2024-03-17     0   1        12
#> 72  2024-03-18    11   4         1
#> 73  2024-03-19    18   4         9
#> 74  2024-03-20    18   0         4
#> 75  2024-03-21    22   0         4
#> 76  2024-03-22    23   0         7
#> 77  2024-03-23    29   0         4
#> 78  2024-03-24    31   0         0
#> 79  2024-03-25    37   0         5
#> 80  2024-03-26    27   2         0
#> 81  2024-03-27    29   1         1
#> 82  2024-03-28    34   0         0
#> 83  2024-03-29    27   0         0
#> 84  2024-03-30    42   4         9
#> 85  2024-03-31    39   0         5
#> 86  2024-04-01    40  10         0
#> 87  2024-04-02    47   0         0
#> 88  2024-04-03    40   4         1
#> 89  2024-04-04    47   0         0
#> 90  2024-04-05    58   4         0
#> 91  2024-04-06    45   0         0
#> 92  2024-04-07    48   7         4
#> 93  2024-04-08    44   5         0
#> 94  2024-04-09    48  10         0
#> 95  2024-04-10    47  11         0
#> 96  2024-04-11    45  19         1
#> 97  2024-04-12    38  23         4
#> 98  2024-04-13    38  26         7
#> 99  2024-04-14    29  21         0
#> 100 2024-04-15    44  35         1
#> 101 2024-04-16    25  25         0
#> 102 2024-04-17    26  29         0
#> 103 2024-04-18    30  30         0
#> 104 2024-04-19    25  28         4
#> 105 2024-04-20    21  34         0
#> 106 2024-04-21    19  39         2
#> 107 2024-04-22    16  42         0
#> 108 2024-04-23    10  33         3
#> 109 2024-04-24    14  41         4
#> 110 2024-04-25    17  27         5
#> 111 2024-04-26     6  40         1
#> 112 2024-04-27    12  41         0
#> 113 2024-04-28    12  46         0
#> 114 2024-04-29     9  39         9
#> 115 2024-04-30     1  40         4
#> 116 2024-05-01     6  28         9
#> 117 2024-05-02     7  26         0
#> 118 2024-05-03     1  28         5
#> 119 2024-05-04     1  35         8
#> 120 2024-05-05     7  24         0
#> 121 2024-05-06     0  10         4
#> 122 2024-05-07     4  32         0
#> 123 2024-05-08     0  14         0
#> 124 2024-05-09     0  22         0
#> 125 2024-05-10     4  12         9
#> 126 2024-05-11     5  18         0
#> 127 2024-05-12     5  14         7
#> 128 2024-05-13     0  10         1
#> 129 2024-05-14     0   4         4
#> 130 2024-05-15     2   6         3
#> 131 2024-05-16     0   4         4
#> 132 2024-05-17     0   0         6
#> 133 2024-05-18     0   8        17
#> 134 2024-05-19     2   0        11
#> 135 2024-05-20     1   5         8
#> 136 2024-05-21     5   2        12
#> 137 2024-05-22     0   3        12
#> 138 2024-05-23     0   3         6
#> 139 2024-05-24     8   0        16
#> 140 2024-05-25     6   0        18
#> 141 2024-05-26     0   5         4
#> 142 2024-05-27     0   2        18
#> 143 2024-05-28     0   0        12
#> 144 2024-05-29     8   0         7
#> 145 2024-05-30     0   3        24
#> 146 2024-05-31     3   7        19
#> 147 2024-06-01     2   1        19
#> 148 2024-06-02     6   5        23
#> 149 2024-06-03     0   0        18
#> 150 2024-06-04     1   0        22
#> 151 2024-06-05     0   0        11
#> 152 2024-06-06     0   6        15
#> 153 2024-06-07     0   3        21
#> 154 2024-06-08     0   0        18
#> 155 2024-06-09     8   6        15
#> 156 2024-06-10     3   2        19
#> 157 2024-06-11     0   9        18
#> 158 2024-06-12     3   0        13
#> 159 2024-06-13     0   0         9
#> 160 2024-06-14     2   0        26
#> 161 2024-06-15     2   0        17
#> 162 2024-06-16     2   4        22
#> 163 2024-06-17     0  10        17
#> 164 2024-06-18     1   0        15
#> 165 2024-06-19     0   0        24
#> 166 2024-06-20     0   5        17
#> 167 2024-06-21     0   0        22
#> 168 2024-06-22     5   0        28
#> 169 2024-06-23     3   0        12
#> 170 2024-06-24     0   9        27
#> 171 2024-06-25     3   0        17
#> 172 2024-06-26     0   1        15
#> 173 2024-06-27     0   0        20
#> 174 2024-06-28     3   0        12
#> 175 2024-06-29     0   0        25
#> 176 2024-06-30     0   0        13
#> 177 2024-07-01     0   2         8
#> 178 2024-07-02     0   1        10
#> 179 2024-07-03     2   8        14
#> 180 2024-07-04     3   7         8
#> 181 2024-07-05     0   0        12
#> 182 2024-07-06     0   1         3
#> 183 2024-07-07     2   3        17
#> 184 2024-07-08     2   0        10
#> 185 2024-07-09     0   0         7
#> 186 2024-07-10     6   3        10
#> 187 2024-07-11     0   0         8
#> 188 2024-07-12     5   0         4
#> 189 2024-07-13     9   0        12
#> 190 2024-07-14     0   0         0
#> 191 2024-07-15     0   0         1
#> 192 2024-07-16     7   0         9
#> 193 2024-07-17     2   0         3
#> 194 2024-07-18     3   0         0
#> 195 2024-07-19     2   6         2
#> 196 2024-07-20     0   0         0
#> 197 2024-07-21     0   1         0
#> 198 2024-07-22     1  12         3
#> 199 2024-07-23     1   0         7
#> 200 2024-07-24     1   3         7
#> 201 2024-07-25     0   0         2
#> 202 2024-07-26     7   0         5
#> 203 2024-07-27     0   3         2
#> 204 2024-07-28     0   0         8
#> 205 2024-07-29     0   0         0
#> 206 2024-07-30     1   0         6
#> 207 2024-07-31     6   0         0
#> 208 2024-08-01     0   0         0
#> 209 2024-08-02     0   0         0
#> 210 2024-08-03     0   0         6
#> 211 2024-08-04     0   2         9
#> 212 2024-08-05    14   0         0
#> 213 2024-08-06     0   1         7
#> 214 2024-08-07    16   0         4
#> 215 2024-08-08     1   0         0
#> 216 2024-08-09     0   2         0
#> 217 2024-08-10     0   4         0
#> 218 2024-08-11     0   3         4
#> 219 2024-08-12     0   5         0
#> 220 2024-08-13     0   0         4
#> 221 2024-08-14     2   7         7
#> 222 2024-08-15     0   0         3
#> 223 2024-08-16     4   0         1
#> 224 2024-08-17     0   1         0
#> 225 2024-08-18     0   0         0
#> 226 2024-08-19     0   6         5
#> 227 2024-08-20     0   0         2
#> 228 2024-08-21     0   1         4
#> 229 2024-08-22     1   3         0
#> 230 2024-08-23     0   0         1
#> 231 2024-08-24     5   0         1
#> 232 2024-08-25     0   4         2
#> 233 2024-08-26     0   3        10
#> 234 2024-08-27     0   6         1
#> 235 2024-08-28     0   0         0
#> 236 2024-08-29     5   4         5
#> 237 2024-08-30     0   0         8
#> 238 2024-08-31     8   7         0
#> 239 2024-09-01     2   0         4
#> 240 2024-09-02     1   0         0
#> 241 2024-09-03     1   0         3
#> 242 2024-09-04     0   3         0
#> 243 2024-09-05    10   0         0
#> 244 2024-09-06     0   0         0
#> 245 2024-09-07     0   1         4
#> 246 2024-09-08     1   0         0
#> 247 2024-09-09     1   0         0
#> 248 2024-09-10     2   0         0
#> 249 2024-09-11     0   0         0
#> 250 2024-09-12     0   0         0
#> 251 2024-09-13     0   0         4
#> 252 2024-09-14     0   4         0
#> 253 2024-09-15     0   0         8
#> 254 2024-09-16     0   7         0
#> 255 2024-09-17     0   0         3
#> 256 2024-09-18     0   0         0
#> 257 2024-09-19     2   6         0
#> 258 2024-09-20     6   1        11
#> 259 2024-09-21     0   4         0
#> 260 2024-09-22     0   0         0
#> 261 2024-09-23     0   0         0
#> 262 2024-09-24     0   0         5
#> 263 2024-09-25     0   0         6
#> 264 2024-09-26     0   0         0
#> 265 2024-09-27     0   1         0
#> 266 2024-09-28     0   0         0
#> 267 2024-09-29     0   5         0
#> 268 2024-09-30     5   2         0
#> 269 2024-10-01     0   0         0
#> 270 2024-10-02     0   0         1
#> 271 2024-10-03     8   0         8
#> 272 2024-10-04     4   0         0
#> 273 2024-10-05     0   7         0
#> 274 2024-10-06     3   0         0
#> 275 2024-10-07     3   0         0
#> 276 2024-10-08     2   3         7
#> 277 2024-10-09     0   0         0
#> 278 2024-10-10     0   0         0
#> 279 2024-10-11     0   0         0
#> 280 2024-10-12     5   4         0
#> 281 2024-10-13     0   0         2
#> 282 2024-10-14     0   3         0
#> 283 2024-10-15     0   3        11
#> 284 2024-10-16     0   0         1
#> 285 2024-10-17     0   5         5
#> 286 2024-10-18     9   6         2
#> 287 2024-10-19     8   0         0
#> 288 2024-10-20     0   0         0
#> 289 2024-10-21     6   0         3
#> 290 2024-10-22     0   0         0
#> 291 2024-10-23     0   8         0
#> 292 2024-10-24     8   0         2
#> 293 2024-10-25     0   5         5
#> 294 2024-10-26     1   0         6
#> 295 2024-10-27     0   0         2
#> 296 2024-10-28     1   5         0
#> 297 2024-10-29     1   3         0
#> 298 2024-10-30     1   3         0
#> 299 2024-10-31     0   0         7
#> 300 2024-11-01     0   7         6
#> 301 2024-11-02     3   0         1
#> 302 2024-11-03    11   0         2
#> 303 2024-11-04     7   4         9
#> 304 2024-11-05     0   0         0
#> 305 2024-11-06     0   1         0
#> 306 2024-11-07     4   0         0
#> 307 2024-11-08     4   0         0
#> 308 2024-11-09     8   2         0
#> 309 2024-11-10     0   7         0
#> 310 2024-11-11     0   4         0
#> 311 2024-11-12     8   0        10
#> 312 2024-11-13    10   1         0
#> 313 2024-11-14     0   0         0
#> 314 2024-11-15     0   0         1
#> 315 2024-11-16    10   0         0
#> 316 2024-11-17     3   0         1
#> 317 2024-11-18     0   0         7
#> 318 2024-11-19     0   0         0
#> 319 2024-11-20     0   1         0
#> 320 2024-11-21     3   0         0
#> 321 2024-11-22     4   2         0
#> 322 2024-11-23     3   0         0
#> 323 2024-11-24     5   0         1
#> 324 2024-11-25     0   7         0
#> 325 2024-11-26     0   5         5
#> 326 2024-11-27     1   0         0
#> 327 2024-11-28     2   0         1
#> 328 2024-11-29     0   0         4
#> 329 2024-11-30     2   0         5
#> 330 2024-12-01     3  11         8
#> 331 2024-12-02     0   0         0
#> 332 2024-12-03     2   0         0
#> 333 2024-12-04     0   0         0
#> 334 2024-12-05     4   0         1
#> 335 2024-12-06     0   0         7
#> 336 2024-12-07     0   0         0
#> 337 2024-12-08     0   0         0
#> 338 2024-12-09     0   3         0
#> 339 2024-12-10     0   6         0
#> 340 2024-12-11     0   0         0
#> 341 2024-12-12     0   6         0
#> 342 2024-12-13     5   0         0
#> 343 2024-12-14     8   0         1
#> 344 2024-12-15     6  12         1
#> 345 2024-12-16     2   0         1
#> 346 2024-12-17     0   0         1
#> 347 2024-12-18     0   4         3
#> 348 2024-12-19     0   0         0
#> 349 2024-12-20     0   8         2
#> 350 2024-12-21     0   0         2
#> 351 2024-12-22     0   0         0
#> 352 2024-12-23     0   1         0
#> 353 2024-12-24     0   2         0
#> 354 2024-12-25     6   4         3
#> 355 2024-12-26     0   0         0
#> 356 2024-12-27    13   2         4
#> 357 2024-12-28     0   6         4
#> 358 2024-12-29     0   0         0
#> 359 2024-12-30     2   0         0
#> 360 2024-12-31     0   0         1
#> 361 2025-01-01     0   0         0
#> 362 2025-01-02     9   6         0
#> 363 2025-01-03    12  10         0
#> 364 2025-01-04     0   0         2
#> 365 2025-01-05     0   1         0
#> 366 2025-01-06     0   0         0
simulate_data(days = 100, peaks = c(flu_a = 30), amplitudes = c(flu_a = 60),
scales = c(flu_a = -0.01), noise_sd = c(flu_a = 5))
#>           date flu_a
#> 1   2024-01-07     0
#> 2   2024-01-08     4
#> 3   2024-01-09     4
#> 4   2024-01-10     0
#> 5   2024-01-11     5
#> 6   2024-01-12     0
#> 7   2024-01-13     0
#> 8   2024-01-14     8
#> 9   2024-01-15    10
#> 10  2024-01-16     0
#> 11  2024-01-17     0
#> 12  2024-01-18     4
#> 13  2024-01-19     0
#> 14  2024-01-20     0
#> 15  2024-01-21    10
#> 16  2024-01-22     0
#> 17  2024-01-23    11
#> 18  2024-01-24     5
#> 19  2024-01-25     8
#> 20  2024-01-26    16
#> 21  2024-01-27    26
#> 22  2024-01-28    13
#> 23  2024-01-29    23
#> 24  2024-01-30    39
#> 25  2024-01-31    36
#> 26  2024-02-01    44
#> 27  2024-02-02    53
#> 28  2024-02-03    57
#> 29  2024-02-04    56
#> 30  2024-02-05    60
#> 31  2024-02-06    54
#> 32  2024-02-07    67
#> 33  2024-02-08    58
#> 34  2024-02-09    48
#> 35  2024-02-10    47
#> 36  2024-02-11    40
#> 37  2024-02-12    35
#> 38  2024-02-13    40
#> 39  2024-02-14    26
#> 40  2024-02-15    24
#> 41  2024-02-16    21
#> 42  2024-02-17    19
#> 43  2024-02-18    15
#> 44  2024-02-19    16
#> 45  2024-02-20     0
#> 46  2024-02-21     0
#> 47  2024-02-22     9
#> 48  2024-02-23     6
#> 49  2024-02-24     1
#> 50  2024-02-25     0
#> 51  2024-02-26     0
#> 52  2024-02-27     0
#> 53  2024-02-28     0
#> 54  2024-02-29     0
#> 55  2024-03-01     0
#> 56  2024-03-02     9
#> 57  2024-03-03     0
#> 58  2024-03-04     0
#> 59  2024-03-05     0
#> 60  2024-03-06     0
#> 61  2024-03-07     3
#> 62  2024-03-08     3
#> 63  2024-03-09     6
#> 64  2024-03-10     4
#> 65  2024-03-11     7
#> 66  2024-03-12     3
#> 67  2024-03-13     0
#> 68  2024-03-14     0
#> 69  2024-03-15     0
#> 70  2024-03-16     0
#> 71  2024-03-17     0
#> 72  2024-03-18     5
#> 73  2024-03-19     0
#> 74  2024-03-20     5
#> 75  2024-03-21     2
#> 76  2024-03-22     0
#> 77  2024-03-23     9
#> 78  2024-03-24     0
#> 79  2024-03-25     7
#> 80  2024-03-26     8
#> 81  2024-03-27     0
#> 82  2024-03-28     0
#> 83  2024-03-29     0
#> 84  2024-03-30     0
#> 85  2024-03-31     0
#> 86  2024-04-01     0
#> 87  2024-04-02     0
#> 88  2024-04-03     0
#> 89  2024-04-04     0
#> 90  2024-04-05     0
#> 91  2024-04-06     0
#> 92  2024-04-07     0
#> 93  2024-04-08     3
#> 94  2024-04-09     0
#> 95  2024-04-10     4
#> 96  2024-04-11     0
#> 97  2024-04-12     0
#> 98  2024-04-13     0
#> 99  2024-04-14     5
#> 100 2024-04-15     0
#> 101 2024-04-16     0